Moved calendar navigation and components
