+++ src/modules/createEvent/Inputs/TimeInput.tsx
@@ -5,7 +5,7 @@ import {
 } from 'CreateEvent/CreateEventFormStore';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import DefaultInputButton from './components/DefaultInputButton';
-import DateTimeLineItem from '../../../components/DateTimeLineItem';
+import DateTimeLineItem from './components/DateTimeLineItem';
 import {View, Platform, Modal} from 'react-native';
 
 const ChooseTime = ({
